feat(examples): add ripple to VueJS example, and reorganize.
* Add a new ripple directive.
* Update the sample app to include a surface with ripple.
* Reorganize things a bit to keep components in distinct
  directories.
