+++ examples/vue/src/v-mdl-checkbox/Checkbox.vue
