+++ examples/vue/src/v-mdl-checkbox/CheckboxLabel.vue
