+++ examples/vue/src/v-mdl-checkbox/CheckboxWrapper.js
